vocore2.dts: declare a partition in spi-cs1
The easiest way to recover a vocore with a wrong boot loader, is
reflashing uboot from another vocore.

To do that, keep the bricked vocore in reset and connect its CS0 to
the CS1 of another device. This partition allows this "other" device
to see the boot loader partition on CS1.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>