mtd/spi-nor:remove non-jedec-id devices
If no flash is there, one of these is actually matched, and we
see an unexistent mtd partition.

I want spi-cs1 to be described in my device tree, so that when a flash
is really there is is used, but I need to not see it in /proc/mtd when
it doesn't exist.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>